Add json account loader
